refactor: FloatingMenu refactoring
