+++ src/domains/menu/views/floating-menu/index.ts
@@ -1,0 +1,1 @@
+export * from './FloatingMenu';
