+++ src/domains/menu/views/menu-base/MenuBase.tsx
@@ -2,17 +2,19 @@ import { cn } from 'third-parties/classnames';
 import { container } from './MenuBase.css';
 import { forwardRef } from 'react';
 
+export type MenuType = 'icon' | 'text' | 'text-bold';
+
 export interface MenuBaseProps extends React.HTMLAttributes<HTMLDivElement> {
   children: React.ReactNode;
-  type: 'apple-logo' | 'app-name' | 'item';
+  type: MenuType;
   className?: string;
   selected: boolean;
 }
 
-function _MenuBase(
+const _MenuBase = (
   { children, type, className, selected, ...rest }: MenuBaseProps,
   ref: React.Ref<HTMLDivElement>
-) {
+) => {
   return (
     <div
       ref={ref}
@@ -22,6 +24,6 @@ function _MenuBase(
       {children}
     </div>
   );
-}
+};
 
 export const MenuBase = forwardRef(_MenuBase);
