--- src/domains/menu/views/menu-left/components/menu-apple-logo/MenuAppleLogo.css.ts
@@ -1,5 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const logoContainer = style({
-  padding: '2px 11px',
-});
