+++ src/domains/menu/views/menu-left/components/menu-item/MenuItem.tsx
@@ -1,11 +1,10 @@
 import { MenuConfig } from 'domains/app/interface';
 import { FloatingMenu } from '../../../floating-menu/FloatingMenu';
 import { SubmenuGroups } from '../submenu-groups/SubMenus';
-import { MenuBase } from 'domains/menu/views/menu-base';
 
 interface MenuItemProps {
   menu: MenuConfig;
-  type: 'app-name' | 'item';
+  type: 'text' | 'text-bold';
   focused: boolean;
   selected: boolean;
   onSelectedChange: (selected: boolean) => void;
@@ -24,11 +23,7 @@ export function MenuItem({
       selected={selected}
       onSelectedChange={onSelectedChange}
     >
-      <FloatingMenu.Trigger type={type}>
-        <MenuBase selected={selected} type={type}>
-          {menu.name}
-        </MenuBase>
-      </FloatingMenu.Trigger>
+      <FloatingMenu.Trigger type={type}>{menu.name}</FloatingMenu.Trigger>
       <FloatingMenu.Content>
         <SubmenuGroups submenuGroups={menu.submenuGroups} />
       </FloatingMenu.Content>
